Fixed texture improper function

+++ OceanExploration/Assets/Scripts/Terrain/CPUMarchingCubes.cs
@@ -50,7 +50,7 @@ public class CPUMarchingCubes {
         0x70c, 0x605, 0x50f, 0x406, 0x30a, 0x203, 0x109, 0x0   };
 
     // Returns the edges of every triangle needed to create the faces. There are 5 possible triangles * 3 edges=15
-    public static sbyte[,] facetsTable = {{-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
+    public static int[,] facetsTable = {{-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
         {0, 8, 3, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
         {0, 1, 9, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
         {1, 8, 3, 9, 8, 1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1},
